Fixed build for esp-idf framework

+++ components/haier/hon_climate.cpp
@@ -302,7 +302,7 @@ void HonClimate::set_handlers() {
 void HonClimate::dump_config() {
   HaierClimateBase::dump_config();
   ESP_LOGCONFIG(TAG, "  Protocol version: hOn");
-  ESP_LOGCONFIG(TAG, "  Control method: %d", this->control_method_);
+  ESP_LOGCONFIG(TAG, "  Control method: %d", (uint8_t)this->control_method_);
   if (this->hvac_hardware_info_available_) {
     ESP_LOGCONFIG(TAG, "  Device protocol version: %s", this->hvac_protocol_version_.c_str());
     ESP_LOGCONFIG(TAG, "  Device software version: %s", this->hvac_software_version_.c_str());
